in progress 2

+++ core/common/src/main/java/com/mychessapp/core/common/numbers/Operator.kt
@@ -2,60 +2,60 @@ package com.mychessapp.core.common.numbers
 
 operator fun Number.minus(other: Number): Number {
   return when (this) {
-    is Long -> this.toLong() - other.toLong()
-    is Int -> this.toInt() - other.toInt()
-    is Short -> this.toShort() - other.toShort()
-    is Byte -> this.toByte() - other.toByte()
-    is Double -> this.toDouble() - other.toDouble()
-    is Float -> this.toFloat() - other.toFloat()
+    is Long -> this - other.toLong()
+    is Int -> this - other.toInt()
+    is Short -> this - other.toShort()
+    is Byte -> this - other.toByte()
+    is Double -> this - other.toDouble()
+    is Float -> this - other.toFloat()
     else -> throw RuntimeException("Unknown numeric type")
   }
 }
 
 operator fun Number.plus(other: Number): Number {
   return when (this) {
-    is Long -> this.toLong() + other.toLong()
-    is Int -> this.toInt() + other.toInt()
-    is Short -> this.toShort() + other.toShort()
-    is Byte -> this.toByte() + other.toByte()
-    is Double -> this.toDouble() + other.toDouble()
-    is Float -> this.toFloat() + other.toFloat()
+    is Long -> this + other.toLong()
+    is Int -> this + other.toInt()
+    is Short -> this + other.toShort()
+    is Byte -> this + other.toByte()
+    is Double -> this + other.toDouble()
+    is Float -> this + other.toFloat()
     else -> throw RuntimeException("Unknown numeric type")
   }
 }
 
 operator fun Number.times(other: Number): Number {
   return when (this) {
-    is Long -> this.toLong() * other.toLong()
-    is Int -> this.toInt() * other.toInt()
-    is Short -> this.toShort() * other.toShort()
-    is Byte -> this.toByte() * other.toByte()
-    is Double -> this.toDouble() * other.toDouble()
-    is Float -> this.toFloat() * other.toFloat()
+    is Long -> this * other.toLong()
+    is Int -> this * other.toInt()
+    is Short -> this * other.toShort()
+    is Byte -> this * other.toByte()
+    is Double -> this * other.toDouble()
+    is Float -> this * other.toFloat()
     else -> throw RuntimeException("Unknown numeric type")
   }
 }
 
 operator fun Number.div(other: Number): Number {
   return when (this) {
-    is Long -> this.toLong() / other.toLong()
-    is Int -> this.toInt() / other.toInt()
-    is Short -> this.toShort() / other.toShort()
-    is Byte -> this.toByte() / other.toByte()
-    is Double -> this.toDouble() / other.toDouble()
-    is Float -> this.toFloat() / other.toFloat()
+    is Long -> this / other.toLong()
+    is Int -> this / other.toInt()
+    is Short -> this / other.toShort()
+    is Byte -> this / other.toByte()
+    is Double -> this / other.toDouble()
+    is Float -> this / other.toFloat()
     else -> throw RuntimeException("Unknown numeric type")
   }
 }
 
 operator fun Number.rem(other: Number): Number {
   return when (this) {
-    is Long -> this.toLong() % other.toLong()
-    is Int -> this.toInt() % other.toInt()
-    is Short -> this.toShort() % other.toShort()
-    is Byte -> this.toByte() % other.toByte()
-    is Double -> this.toDouble() % other.toDouble()
-    is Float -> this.toFloat() % other.toFloat()
+    is Long -> this % other.toLong()
+    is Int -> this % other.toInt()
+    is Short -> this % other.toShort()
+    is Byte -> this % other.toByte()
+    is Double -> this % other.toDouble()
+    is Float -> this % other.toFloat()
     else -> throw RuntimeException("Unknown numeric type")
   }
 }
